Conflicts:
	MerlionBank-ejb/src/java/CommonEntity/Staff.java
	MerlionBank-ejb/src/java/CustomerRelationshipEntity/CaseEntity.java
	MerlionBank-ejb/src/java/CustomerRelationshipEntity/Issue.java
	MerlionBank-ejb/src/java/CustomerRelationshipEntity/Session/CollaborativeCRMSessionBean.java

+++ MerlionBank-ejb/src/java/CustomerRelationshipEntity/Issue.java
@@ -1,3 +1,4 @@
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
@@ -40,7 +41,7 @@ public class Issue implements Serializable {
     private String content;
     private String issueType; 
     private String status;
-    private Double rating;
+    private int rating;
     private String solution;
   
     @ManyToOne
@@ -164,14 +165,14 @@ public class Issue implements Serializable {
     /**
      * @return the rating
      */
-    public Double getRating() {
+    public int getRating() {
         return rating;
     }
 
     /**
      * @param rating the rating to set
      */
-    public void setRating(Double rating) {
+    public void setRating(int rating) {
         this.rating = rating;
     }
